Add feedback updating using an internal command queue
The PTZ Optics camera command API responds to most commands with an ACK response
followed by a 'Complete' response. These are helpful for knowing the state of the last command.
However, they do not include what the last command was that it executed. It assumes that
the software keeps track of which commands are sent to it. This necessitated adding logic to
keep track of commands that are sent to the camera. Along with that, we have to make sure
we keep our internal command queue in sync with the actual state of the camera. This can
be tricky with TCP errors or other errors that the camera returns back. We've tried to account
for all these scenarios and update the command queue in an appropriate way.

Whenever there is a new command added to the module, they must follow the example
of the other commands that push the command onto the queue. If not, the queue will
quickly get out of sync.

There are a set of commands to the camera called Inquiry commands that allow you to query
the camera for certain state. These commands do not follow the ACK + 'Complete'. This commit
allows for logging these responses for informational purposes. It also sets it up for adding
specific logic to update future feedbacks based on these responses.